Cleaning up deps and adding abstractions.
