+++ package.json
@@ -30,9 +30,7 @@
     "cashaccounts": "0.4.0",
     "cookie-parser": "1.4.4",
     "cors": "2.8.5",
-    "debug": "~2.6.3",
-    "elasticsearch": "^13.0.1",
-    "express": "^4.15.5",
+    "express": "4.17.1",
     "express-basic-auth": "^1.1.3",
     "express-rate-limit": "^2.11.0",
     "gulp-json-modify": "^1.0.2",
